requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
anthropic==0.40.0
gitpython==3.1.43
PyGithub==2.4.0
httpx==0.27.2
